refactor unicorn background
